requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
python-telegram-bot==21.7
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.3